backend/pyproject.toml
# pyproject.toml
# ──────────────────────────────────────────────
# AH_02_07 HealthGuide 프로젝트 루트 설정
# uv sync 로 가상환경 생성 + 의존성 설치
# ──────────────────────────────────────────────

[project]
name = "healthguide"
version = "0.1.0"
description = "AI 기반 헬스케어 상담 웹 서비스"
requires-python = ">=3.13"

# ── 공통 의존성 (API 서버 + Worker 모두 사용) ──
dependencies = [
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "redis>=5.2",
    "python-dotenv>=1.0",
]

# ── 의존성 그룹: uv sync --group <name> 으로 선택 설치 ──
[project.optional-dependencies]
app = [
    # FastAPI 서버 관련
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "tortoise-orm[asyncpg]>=0.24",
    "aerich>=0.8",             # Tortoise ORM 마이그레이션
    "asyncpg>=0.30",
    "sse-starlette>=2.2",      # Server-Sent Events
    "python-multipart>=0.0.18",# 파일 업로드
    "python-jose[cryptography]>=3.3",  # JWT
    "passlib[bcrypt]>=1.7",    # 비밀번호 해싱
    "httpx>=0.28",             # 외부 API 호출 (OAuth 등)
    "boto3>=1.36",             # AWS S3
    "openai>=1.60",            # OpenAI GPT API
    "email-validator>=2.0",
]
ai = [
    # AI Worker 관련
    "torch>=2.6",
    "vllm>=0.7",
    "transformers>=4.48",
    "Pillow>=11.1",            # 이미지 처리 (알약 분석)
    "boto3>=1.36",             # S3에서 모델/이미지 다운로드
    "openai>=1.60",
]
dev = [
    # 개발 도구
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "httpx>=0.28",             # 테스트 클라이언트
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "ai_worker"]

# ── Ruff (린터 + 포매터) 설정 ──
[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]  # line-length는 formatter가 처리

[tool.ruff.format]
quote-style = "double"

# ── Mypy 설정 ──
[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

# ── Pytest 설정 ──
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

# ── Aerich (Tortoise ORM 마이그레이션) 설정 ──
[tool.aerich]
tortoise_orm = "app.db.databases.TORTOISE_ORM"
location = "./migrations"
src_folder = "."
